Fix MOI used in recovery and windup time calcs.
Specifically:

Multiply torque by motor ratio. Note that doing this, combined with the
existing division of the freespeed by the motor ratio, is equivalent to
dividing the moi by the motor ratio squared.

Update tests to reflect new expected results.

Calculate windup time as recovery time from a speed of zero to eliminate
duplication of calculation.

Rename a couple variable for clarity.