Fix calculation of Ka.
Specifically:

Adjust MOI by the motor ratio squared to get MOI seen by motor.

Divide that MOI by square of radius of shooter instead of flywheel since
the surface acceleration of the shooter is  what the kA will be for.

Don't multiply the motor torque by the motor ratio since that is now
addressed by using the modified MOI.